[cef] use CefResourceRequestHandler to handle HTTP requests
Was introduced anyway for batch job handling, where no THttpServer
instance is created. Let avoid custom singelton with scheme handler, 
which only allowed to use the only THttpServer instance
