+++ gui/cefdisplay/src/gui_handler.cxx
@@ -19,6 +19,7 @@
 #endif
 
 #include "gui_handler.h"
+#include "simple_app.h"
 
 #include <sstream>
 #include <string>
@@ -31,13 +32,18 @@
 #include "include/wrapper/cef_closure_task.h"
 #include "include/wrapper/cef_helpers.h"
 #include "include/cef_parser.h"
-
+#include "include/wrapper/cef_stream_resource_handler.h"
 
 #include "TEnv.h"
+#include "TUrl.h"
+#include "THttpServer.h"
+#include "THttpCallArg.h"
+#include "TSystem.h"
+#include "TBase64.h"
 #include <ROOT/RLogger.hxx>
 
 
-GuiHandler::GuiHandler(THttpServer *serv, bool use_views) : fServer(serv), fUseViews(use_views), is_closing_(false)
+GuiHandler::GuiHandler(bool use_views) : fUseViews(use_views), is_closing_(false)
 {
    fConsole = gEnv->GetValue("WebGui.Console", (int)0);
 
@@ -68,7 +74,7 @@ void GuiHandler::OnAfterCreated(CefRefPtr<CefBrowser> browser)
    CEF_REQUIRE_UI_THREAD();
 
    // Add to the list of existing browsers.
-   browser_list_.push_back(browser);
+   fBrowserList.emplace_back(browser);
 }
 
 bool GuiHandler::DoClose(CefRefPtr<CefBrowser> browser)
@@ -78,7 +84,7 @@ bool GuiHandler::DoClose(CefRefPtr<CefBrowser> browser)
    // Closing the main window requires special handling. See the DoClose()
    // documentation in the CEF header for a detailed description of this
    // process.
-   if (browser_list_.size() == 1) {
+   if (fBrowserList.size() == 1) {
       // Set a flag to indicate that the window close should be allowed.
       is_closing_ = true;
    }
@@ -93,15 +99,15 @@ void GuiHandler::OnBeforeClose(CefRefPtr<CefBrowser> browser)
    CEF_REQUIRE_UI_THREAD();
 
    // Remove from the list of existing browsers.
-   BrowserList::iterator bit = browser_list_.begin();
-   for (; bit != browser_list_.end(); ++bit) {
+   auto bit = fBrowserList.begin();
+   for (; bit != fBrowserList.end(); ++bit) {
       if ((*bit)->IsSame(browser)) {
-         browser_list_.erase(bit);
+         fBrowserList.erase(bit);
          break;
       }
    }
 
-   if (browser_list_.empty()) {
+   if (fBrowserList.empty()) {
 
       // All browser windows have closed. Quit the application message loop.
 
@@ -146,12 +152,11 @@ void GuiHandler::CloseAllBrowsers(bool force_close)
       return;
    }
 
-   if (browser_list_.empty())
+   if (fBrowserList.empty())
       return;
 
-   BrowserList::const_iterator it = browser_list_.begin();
-   for (; it != browser_list_.end(); ++it)
-      (*it)->GetHost()->CloseBrowser(force_close);
+   for (auto &br : fBrowserList)
+      br->GetHost()->CloseBrowser(force_close);
 }
 
 bool GuiHandler::OnConsoleMessage(CefRefPtr<CefBrowser> browser,
@@ -193,17 +198,236 @@ cef_return_value_t GuiHandler::OnBeforeResourceLoad(
                                                  callback);
 }
 
+
+
+class TCefHttpCallArg : public THttpCallArg {
+protected:
+
+   CefRefPtr<CefCallback> fCallBack{nullptr};
+
+   void CheckWSPageContent(THttpWSHandler *) override
+   {
+      std::string search = "JSROOT.ConnectWebWindow({";
+      std::string replace = search + "platform:\"cef3\",socket_kind:\"longpoll\",";
+
+      ReplaceAllinContent(search, replace, true);
+   }
+
+public:
+   explicit TCefHttpCallArg() = default;
+
+   void AssignCallback(CefRefPtr<CefCallback> cb) { fCallBack = cb; }
+
+   // this is callback from HTTP server
+   void HttpReplied() override
+   {
+      if (IsFile()) {
+         // send file
+         std::string file_content = THttpServer::ReadFileContent((const char *)GetContent());
+         SetContent(std::move(file_content));
+      }
+
+      fCallBack->Continue(); // we have result and can continue with processing
+   }
+};
+
+
+class TGuiResourceHandler : public CefResourceHandler {
+public:
+   // QWebEngineUrlRequestJob *fRequest;
+
+   THttpServer *fServer{nullptr};
+   std::shared_ptr<TCefHttpCallArg> fArg;
+
+   int fTransferOffset{0};
+
+   explicit TGuiResourceHandler(THttpServer *serv, bool dummy = false)
+   {
+      fServer = serv;
+
+      if (!dummy)
+         fArg = std::make_shared<TCefHttpCallArg>();
+   }
+
+   virtual ~TGuiResourceHandler() {}
+
+   void Cancel() OVERRIDE { CEF_REQUIRE_IO_THREAD(); }
+
+   bool ProcessRequest(CefRefPtr<CefRequest> request, CefRefPtr<CefCallback> callback) OVERRIDE
+   {
+      CEF_REQUIRE_IO_THREAD();
+
+      if (fArg) {
+         fArg->AssignCallback(callback);
+         fServer->SubmitHttp(fArg);
+      } else {
+         callback->Continue();
+      }
+
+      return true;
+   }
+
+   void GetResponseHeaders(CefRefPtr<CefResponse> response, int64 &response_length, CefString &redirectUrl) OVERRIDE
+   {
+      CEF_REQUIRE_IO_THREAD();
+
+      if (!fArg || fArg->Is404()) {
+         response->SetMimeType("text/html");
+         response->SetStatus(404);
+         response_length = 0;
+      } else {
+         response->SetMimeType(fArg->GetContentType());
+         response->SetStatus(200);
+         response_length = fArg->GetContentLength();
+
+         if (fArg->NumHeader() > 0) {
+            // printf("******* Response with extra headers\n");
+            CefResponse::HeaderMap headers;
+            for (Int_t n = 0; n < fArg->NumHeader(); ++n) {
+               TString name = fArg->GetHeaderName(n);
+               TString value = fArg->GetHeader(name.Data());
+               headers.emplace(CefString(name.Data()), CefString(value.Data()));
+               // printf("   header %s %s\n", name.Data(), value.Data());
+            }
+            response->SetHeaderMap(headers);
+         }
+//         if (strstr(fArg->GetQuery(),"connection="))
+//            printf("Reply %s %s %s  len: %d %s\n", fArg->GetPathName(), fArg->GetFileName(), fArg->GetQuery(),
+//                  fArg->GetContentLength(), (const char *) fArg->GetContent() );
+      }
+      // DCHECK(!fArg->Is404());
+   }
+
+   bool ReadResponse(void *data_out, int bytes_to_read, int &bytes_read, CefRefPtr<CefCallback> callback) OVERRIDE
+   {
+      CEF_REQUIRE_IO_THREAD();
+
+      if (!fArg) return false;
+
+      bytes_read = 0;
+
+      if (fTransferOffset < fArg->GetContentLength()) {
+         char *data_ = (char *)fArg->GetContent();
+         // Copy the next block of data into the buffer.
+         int transfer_size = fArg->GetContentLength() - fTransferOffset;
+         if (transfer_size > bytes_to_read)
+            transfer_size = bytes_to_read;
+         memcpy(data_out, data_ + fTransferOffset, transfer_size);
+         fTransferOffset += transfer_size;
+
+         bytes_read = transfer_size;
+      }
+
+      // if content fully copied - can release reference, object will be cleaned up
+      if (fTransferOffset >= fArg->GetContentLength())
+         fArg.reset();
+
+      return bytes_read > 0;
+   }
+
+   IMPLEMENT_REFCOUNTING(TGuiResourceHandler);
+   DISALLOW_COPY_AND_ASSIGN(TGuiResourceHandler);
+};
+
+
 CefRefPtr<CefResourceHandler> GuiHandler::GetResourceHandler(
     CefRefPtr<CefBrowser> browser,
     CefRefPtr<CefFrame> frame,
     CefRefPtr<CefRequest> request) {
   CEF_REQUIRE_IO_THREAD();
 
-  // std::string url = request->GetURL().ToString();
-  // printf("GetResourceHandler url %s\n", url.c_str());
+  std::string addr = request->GetURL().ToString();
+  std::string prefix = "http://rootserver.local";
 
-  return fResourceManager->GetResourceHandler(browser, frame, request);
+  if (addr.compare(0, prefix.length(), prefix) != 0)
+     return fResourceManager->GetResourceHandler(browser, frame, request);
+
+  int indx = std::stoi(addr.substr(prefix.length(), addr.find("/", prefix.length()) - prefix.length()));
+
+  if ((indx < 0) || (indx >= (int) fServers.size()) || !fServers[indx]) {
+     R__ERROR_HERE("CEF") << "No THttpServer with index " << indx;
+     return nullptr;
+  }
+
+  THttpServer *serv = fServers[indx];
+  if (serv->IsZombie()) {
+     fServers[indx] = nullptr;
+     R__ERROR_HERE("CEF") << "THttpServer with index " << indx << " is zombie now";
+     return nullptr;
+  }
+
+  TUrl url(addr.c_str());
+
+  const char *inp_path = url.GetFile();
+
+  TString inp_query = url.GetOptions();
+
+  TString fname;
+
+  if (serv->IsFileRequested(inp_path, fname)) {
+     // process file - no need for special requests handling
+
+     // when file not exists - return nullptr
+     if (gSystem->AccessPathName(fname.Data()))
+        return new TGuiResourceHandler(serv, true);
+
+     const char *mime = THttpServer::GetMimeType(fname.Data());
+
+     CefRefPtr<CefStreamReader> stream = CefStreamReader::CreateForFile(fname.Data());
+
+     // Constructor for HTTP status code 200 and no custom response headers.
+     // There’s also a version of the constructor for custom status code and response headers.
+     return new CefStreamResourceHandler(mime, stream);
+  }
+
+  std::string inp_method = request->GetMethod().ToString();
+
+  // printf("REQUEST METHOD %s\n", inp_method.c_str());
+
+  TGuiResourceHandler *handler = new TGuiResourceHandler(serv);
+  handler->fArg->SetMethod(inp_method.c_str());
+  handler->fArg->SetPathAndFileName(inp_path);
+  handler->fArg->SetTopName("webgui");
+
+  if (inp_method == "POST") {
+
+     CefRefPtr< CefPostData > post_data = request->GetPostData();
+
+     if (!post_data) {
+        R__ERROR_HERE("CEF") << "FATAL - NO POST DATA in CEF HANDLER!!!";
+        exit(1);
+     } else {
+        CefPostData::ElementVector elements;
+        post_data->GetElements(elements);
+        size_t sz = 0, off = 0;
+        for (unsigned n = 0; n < elements.size(); ++n)
+           sz += elements[n]->GetBytesCount();
+        std::string data;
+        data.resize(sz);
+
+        for (unsigned n = 0; n < elements.size(); ++n) {
+           sz = elements[n]->GetBytes(elements[n]->GetBytesCount(), (char *)data.data() + off);
+           off += sz;
+        }
+        handler->fArg->SetPostData(std::move(data));
+     }
+  } else if (inp_query.Index("&post=") != kNPOS) {
+     Int_t pos = inp_query.Index("&post=");
+     TString buf = TBase64::Decode(inp_query.Data() + pos + 6);
+     handler->fArg->SetPostData(std::string(buf.Data()));
+     inp_query.Resize(pos);
+  }
+
+  handler->fArg->SetQuery(inp_query.Data());
+
+  // just return handler
+  return handler;
 }
+
+///////////////////////////////////////////////////////////////////////////////////////////////////
+/// Generate URL for batch page
+/// Uses file:/// prefix to let access JSROOT scripts placed on file system
+/// Register provider for that page in resource manager
 
 std::string GuiHandler::AddBatchPage(const std::string &cont)
 {
@@ -214,3 +438,21 @@ std::string GuiHandler::AddBatchPage(const std::string &cont)
    fResourceManager->AddContentProvider(url, cont, "text/html", 0, std::string());
    return url;
 }
+
+///////////////////////////////////////////////////////////////////////////////////////////////////
+/// Generate URL for RWebWindow page
+/// Register server instance and assign it with the index
+/// Produced URL only works inside CEF and does not represent real HTTP address
+
+std::string GuiHandler::MakePageUrl(THttpServer *serv, const std::string &addr)
+{
+   unsigned indx = 0;
+   while ((indx < fServers.size()) && (fServers[indx] != serv)) indx++;
+   if (indx >= fServers.size()) {
+      fServers.emplace_back(serv);
+      indx = fServers.size() - 1;
+   }
+
+   return std::string("http://rootserver.local") + std::to_string(indx) + addr;
+}
+
